部分协议类添加继承AbstractMessage

+++ core/src/main/java/com/jjg/game/core/pb/ReqBuyProduct.java
@@ -14,7 +14,7 @@ import com.jjg.game.core.constant.ShopConstant;
 @ProtoDesc("商城中物品兑换")
 public class ReqBuyProduct extends AbstractMessage {
     @ProtoDesc("商品id")
-    public int productId;
+    public long productId;
     @ProtoDesc("兑换个数")
     public int count;
 }
